# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Role-Based Approval & Workflow Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List all requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: submittedBy
          in: query
          schema:
            type: integer
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestList"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a new request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
  /requests/{requestId}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get request by ID
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request details with approval steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "404":
          description: Not found
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
    delete:
      operationId: deleteRequest
      tags: [requests]
      summary: Delete request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /requests/{requestId}/approve:
    post:
      operationId: approveRequest
      tags: [requests]
      summary: Approve a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalAction"
      responses:
        "200":
          description: Approval step updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalStep"
  /requests/{requestId}/reject:
    post:
      operationId: rejectRequest
      tags: [requests]
      summary: Reject a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalAction"
      responses:
        "200":
          description: Approval step updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalStep"

  # Workflow Templates
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, department, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, manager, employee]
        department:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Request:
      type: object
      required: [id, title, type, status, priority, submittedById, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [leave, expense, procurement, it_access, hr_policy, other]
        status:
          type: string
          enum: [draft, pending, in_review, approved, rejected, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        submittedById:
          type: integer
        submittedBy:
          $ref: "#/components/schemas/User"
        assignedToId:
          type: ["integer", "null"]
        assignedTo:
          anyOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        workflowTemplateId:
          type: ["integer", "null"]
        currentStepOrder:
          type: ["integer", "null"]
        totalSteps:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    RequestDetail:
      allOf:
        - $ref: "#/components/schemas/Request"
        - type: object
          properties:
            approvalSteps:
              type: array
              items:
                $ref: "#/components/schemas/ApprovalStep"
            comments:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
            auditLog:
              type: array
              items:
                $ref: "#/components/schemas/AuditLogEntry"
    RequestList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Request"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    RequestInput:
      type: object
      required: [title, type, priority]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        type:
          type: string
          enum: [leave, expense, procurement, it_access, hr_policy, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        workflowTemplateId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
    RequestUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [leave, expense, procurement, it_access, hr_policy, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [draft, pending, in_review, approved, rejected, cancelled]
        assignedToId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]

    # Approval
    ApprovalStep:
      type: object
      required: [id, requestId, stepOrder, approverRole, status, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        stepOrder:
          type: integer
        approverRole:
          type: string
        approverId:
          type: ["integer", "null"]
        approver:
          anyOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        status:
          type: string
          enum: [pending, approved, rejected, skipped]
        comment:
          type: ["string", "null"]
        decidedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ApprovalAction:
      type: object
      properties:
        comment:
          type: string

    # Workflow Template
...
    Comment:
      type: object
      required: [id, requestId, userId, text, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        text:
          type: string
        createdAt:
          type: string
...
    AuditLogEntry:
      type: object
      required: [id, action, performedById, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: ["integer", "null"]
        action:
          type: string
        performedById:
          type: integer
        performedBy:
          $ref: "#/components/schemas/User"
        details:
          type: ["string", "null"]
        createdAt:
          type: string
...