# lib/api-zod/src/generated/types/requestDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Role-Based Approval & Workflow Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalStep } from './approvalStep';
import type { AuditLogEntry } from './auditLogEntry';
import type { Comment } from './comment';
import type { Request } from './request';

export type RequestDetail = Request & {
  approvalSteps?: ApprovalStep[];
  comments?: Comment[];
  auditLog?: AuditLogEntry[];
};